test(select): migrate fill
