+++ core/src/components/select/test/fill/select.e2e.ts
@@ -1,0 +1,169 @@
+import { expect } from '@playwright/test';
+import { configs, test } from '@utils/test/playwright';
+
+configs({ modes: ['md'] }).forEach(({ title, screenshot, config }) => {
+  test.describe(title('select: fill'), () => {
+    test.describe('select: fill solid', () => {
+      test('should not have visual regressions', async ({ page }) => {
+        await page.setContent(
+          `
+          <ion-select
+            fill="solid"
+            label="Fruit"
+            value="apple"
+          >
+            <ion-select-option value="apple">Apple</ion-select-option>
+          </ion-select>
+        `,
+          config
+        );
+
+        const select = page.locator('ion-select');
+        expect(await select.screenshot()).toMatchSnapshot(screenshot(`select-fill-solid`));
+      });
+      test('should render correctly with floating label', async ({ page }) => {
+        await page.setContent(
+          `
+          <ion-select
+            fill="solid"
+            label="Fruit"
+            label-placement="floating"
+            value="apple"
+          >
+            <ion-select-option value="apple">Apple</ion-select-option>
+          </ion-select>
+        `,
+          config
+        );
+
+        const select = page.locator('ion-select');
+        expect(await select.screenshot()).toMatchSnapshot(screenshot(`select-fill-solid-label-floating`));
+      });
+      test('should not have visual regressions with shaped solid', async ({ page }) => {
+        await page.setContent(
+          `
+          <ion-select
+            shape="round"
+            fill="solid"
+            label="Fruit"
+            value="apple"
+          >
+            <ion-select-option value="apple">Apple</ion-select-option>
+          </ion-select>
+        `,
+          config
+        );
+
+        const select = page.locator('ion-select');
+        expect(await select.screenshot()).toMatchSnapshot(screenshot(`select-fill-shaped-solid`));
+      });
+      test('padding and border radius should be customizable', async ({ page }) => {
+        await page.setContent(
+          `
+          <style>
+            ion-select {
+              --border-radius: 10px !important;
+              --padding-start: 50px !important;
+              --padding-end: 50px !important;
+            }
+          </style>
+          <ion-select
+            shape="round"
+            fill="solid"
+            label="Fruit"
+            label-placement="floating"
+            value="apple"
+          >
+            <ion-select-option value="apple">Apple</ion-select-option>
+          </ion-select>
+        `,
+          config
+        );
+
+        const select = page.locator('ion-select');
+        expect(await select.screenshot()).toMatchSnapshot(screenshot(`select-fill-shaped-solid-custom`));
+      });
+    });
+    test.describe('select: fill outline', () => {
+      test('should not have visual regressions', async ({ page }) => {
+        await page.setContent(
+          `
+          <ion-select
+            fill="outline"
+            label="Fruit"
+            value="apple"
+          >
+            <ion-select-option value="apple">Apple</ion-select-option>
+          </ion-select>
+        `,
+          config
+        );
+
+        const select = page.locator('ion-select');
+        expect(await select.screenshot()).toMatchSnapshot(screenshot(`select-fill-outline`));
+      });
+      test('should render correctly with floating label', async ({ page }) => {
+        await page.setContent(
+          `
+          <ion-select
+            fill="outline"
+            label="Fruit"
+            label-placement="floating"
+            value="apple"
+          >
+            <ion-select-option value="apple">Apple</ion-select-option>
+          </ion-select>
+        `,
+          config
+        );
+
+        const select = page.locator('ion-select');
+        expect(await select.screenshot()).toMatchSnapshot(screenshot(`select-fill-outline-label-floating`));
+      });
+      test('should not have visual regressions with shaped outline', async ({ page }) => {
+        await page.setContent(
+          `
+          <ion-select
+            shape="round"
+            fill="outline"
+            label="Fruit"
+            value="apple"
+          >
+            <ion-select-option value="apple">Apple</ion-select-option>
+          </ion-select>
+        `,
+          config
+        );
+
+        const select = page.locator('ion-select');
+        expect(await select.screenshot()).toMatchSnapshot(screenshot(`select-fill-shaped-outline`));
+      });
+      test('padding and border radius should be customizable', async ({ page }) => {
+        await page.setContent(
+          `
+          <style>
+            ion-select {
+              --border-radius: 10px !important;
+              --padding-start: 50px !important;
+              --padding-end: 50px !important;
+            }
+          </style>
+          <ion-select
+            shape="round"
+            fill="outline"
+            label="Fruit"
+            label-placement="floating"
+            value="apple"
+          >
+            <ion-select-option value="apple">Apple</ion-select-option>
+          </ion-select>
+        `,
+          config
+        );
+
+        const select = page.locator('ion-select');
+        expect(await select.screenshot()).toMatchSnapshot(screenshot(`select-fill-shaped-outline-custom`));
+      });
+    });
+  });
+});
