+++ core/src/components/select/test/fill/select.e2e.ts-snapshots/select-fill-outline-label-floating-md-ltr-Mobile-Chrome-linux.png
