+++ core/src/components/select/test/fill/select.e2e.ts-snapshots/select-fill-outline-label-floating-md-rtl-Mobile-Firefox-linux.png
