+++ core/src/components/select/test/fill/select.e2e.ts-snapshots/select-fill-outline-md-ltr-Mobile-Chrome-linux.png
