+++ core/src/components/select/test/fill/select.e2e.ts-snapshots/select-fill-shaped-outline-custom-md-rtl-Mobile-Safari-linux.png
