+++ core/src/components/select/test/fill/select.e2e.ts-snapshots/select-fill-solid-label-floating-md-ltr-Mobile-Firefox-linux.png
