+++ core/src/components/select/test/fill/select.e2e.ts-snapshots/select-fill-solid-label-floating-md-rtl-Mobile-Safari-linux.png
