+++ core/src/components/select/test/fill/select.e2e.ts-snapshots/select-fill-solid-md-ltr-Mobile-Chrome-linux.png
