+++ core/src/components/select/test/fill/select.e2e.ts-snapshots/select-fill-solid-md-rtl-Mobile-Chrome-linux.png
